requirements.txt
streamlit
tensorflow
opencv-python-headless
pillow
numpy
h5py